feat: Implement RESTful API for InfoTanques app
The recent code changes include updating the Django project and InfoTanques app to implement a RESTful API. This allows for better communication between the frontend and backend. The changes also include creating the React project and completing the minimum backend functionality. The API can now be executed and viewed locally in the browser.

Based on the recent user commits and repository commits, it seems that the focus of the changes is on updating and implementing the RESTful API. Therefore, the suggested commit message is "feat: Implement RESTful API for InfoTanques app".